Wired selection action for categoryCollectionView with HomeViewController Navigated to RecipesTableViewDetails
