+++ COOKBOOK/Controllers/Home/HomeViewController.swift
@@ -12,11 +12,7 @@ import Kingfisher
 import SideMenu
 
 class HomeViewController: UIViewController {
-    
-    var recipes: Recipes?
-    var recipesDetails = [Recipe]()
-    let indicator = ActivityIndicator()
-    
+
     let searchController = UISearchController(searchResultsController: nil)
     let leftMenuNavigationController = SideMenuNavigationController(rootViewController: SideMenuTableViewController())
         
@@ -92,39 +88,13 @@ extension HomeViewController: UISearchControllerDelegate, UISearchBarDelegate {
     @objc func menuButtonTapped() {
         self.present(leftMenuNavigationController, animated: true, completion: nil)
     }
-    
-    func fetchData(_ category: String) {
-        indicator.setupIndicatorView(view, containerColor: .customDarkGray(), indicatorColor: .white)
-        AF.request("https://api.spoonacular.com/recipes/random?apiKey=e8d1d8ab81044fe491a35c7d370eb5ce&number=25&tags=\(category)").responseJSON { (response) in
-            if let error = response.error {
-                print(error)
-            }
-            do {
-                if let data = response.data {
-                    self.recipes = try JSONDecoder().decode(Recipes.self, from: data)
-                    self.recipesDetails = self.recipes?.recipes ?? []
-                    DispatchQueue.main.async {
-                        self.mainView.foodTableView.reloadData()
-                    }
-                }
-                
-            } catch {
-                print(error)
-            }
-            self.indicator.hideIndicatorView(self.view)
-        }
-    }
-    
 }
 
 extension HomeViewController: HomeViewDidSelectActionDelegate {
     
-    func recipesSelectionAction(indexPath: IndexPath) {
+    func homeView(_ view: HomeView, didSelectCategoryWithTitle title: String) {
         let vc = RecipesTableViewDetails()
-//        vc.mainView.fetchData(category)
-//        fetchData(category)
-//        vc.mainView.foodTableView.reloadData()
+        vc.categoryTitle = title
         self.show(vc, sender: nil)
     }
-
 }
